+++ packages/ui/src/components/tools/items/ToolsCustomItem.jsx
@@ -3,7 +3,11 @@ import { ToolsItem } from "./ToolsItem";
 
 export class ToolsCustomItem extends ToolsItem {
   template() {
-    return <div class="elf--tools-item custom">{this.props.render?.()}</div>;
+    return (
+      <div class="elf--tools-item custom hoverable">
+        {this.props.render?.()}
+      </div>
+    );
   }
 }