+++ apps/editor-dev/CHANGELOG.md
@@ -1,5 +1,11 @@
 # editor-dev
 
+## 1.0.6
+
+### Patch Changes
+
+- update menu items
+
 ## 1.0.5
 
 ### Patch Changes